Move prediction commands out of mod commands
12+ month T3 subs can now run predictions, so moving the commands out of the mod domain makes sense. This also allows better in-discord slash command permission management.
Furthermore, this increases the capacity for slash commands in the mod domain by the 5 we remove
