+++ commands/mod_commands.py
@@ -41,12 +41,9 @@ LOG = logging.getLogger(__name__)
 JOEL_DISCORD_ID = 112386674155122688
 HOOJ_DISCORD_ID = 82969926125490176
 POINTS_AUDIT_CHANNEL = Config.CONFIG["Discord"]["ChannelPoints"]["PointsAuditChannel"]
-PREDICTION_AUDIT_CHANNEL = Config.CONFIG["Discord"]["Predictions"]["AuditChannel"]
 TIER1_ROLE = Config.CONFIG["Discord"]["Subscribers"]["Tier1Role"]
 TIER2_ROLE = Config.CONFIG["Discord"]["Subscribers"]["Tier2Role"]
 TIER3_ROLE = Config.CONFIG["Discord"]["Subscribers"]["Tier3Role"]
-TIER3_ROLE_12MO = Config.CONFIG["Discord"]["Subscribers"]["12MonthTier3Role"]
-CHAT_MOD_ROLE = Config.CONFIG["Discord"]["Roles"]["CMChatModerator"]
 BOT_ROLE = Config.CONFIG["Discord"]["Roles"]["Bot"]
 GIFTED_TIER1_ROLE = Config.CONFIG["Discord"]["Subscribers"]["GiftedTier1Role"]
 GIFTED_TIER2_ROLE = Config.CONFIG["Discord"]["Subscribers"]["GiftedTier3Role"]
@@ -292,68 +289,6 @@ class ModCommands(app_commands.Group, name="mod"):
         await interaction.response.send_message(
             f"Redemptions are currently {status_message}.", ephemeral=True
         )
-
-    @app_commands.command(name="start_prediction")
-    @app_commands.checks.has_any_role(TIER3_ROLE_12MO, MOD_ROLE, CHAT_MOD_ROLE)
-    @app_commands.describe(
-        set_nickname="Whether to prepend users names with their choice"
-    )
-    async def start_prediction(
-        self, interaction: Interaction, set_nickname: Optional[bool] = False
-    ):
-        """Start new prediction"""
-        if DB().has_ongoing_prediction(interaction.guild_id):
-            return await interaction.response.send_message(
-                "There is already an ongoing prediction!", ephemeral=True
-            )
-        await interaction.response.send_modal(
-            CreatePredictionModal(self.client, set_nickname)
-        )
-
-    @app_commands.command(name="refund_prediction")
-    @app_commands.checks.has_role(MOD_ROLE)
-    async def refund_prediction(self, interaction: Interaction):
-        """Refund ongoing prediction, giving users back the points they wagered"""
-        await PayoutPredictionController.refund_prediction(interaction, self.client)
-
-    @app_commands.command(name="close_prediction")
-    @app_commands.checks.has_any_role(TIER3_ROLE_12MO, MOD_ROLE, CHAT_MOD_ROLE)
-    async def close_prediction(self, interaction: Interaction):
-        """CLOSE PREDICTION"""
-        await ClosePredictionController.close_prediction(interaction.guild_id)
-        prediction_id = DB().get_ongoing_prediction_id(interaction.guild_id)
-        prediction_message_id = DB().get_prediction_message_id(prediction_id)
-        prediction_channel_id = DB().get_prediction_channel_id(prediction_id)
-        prediction_message = await self.client.get_channel(
-            prediction_channel_id
-        ).fetch_message(prediction_message_id)
-
-        audit_channel = interaction.guild.get_channel(PREDICTION_AUDIT_CHANNEL)
-        await audit_channel.send(
-            f"{interaction.user.mention} closed the current prediction.",
-            allowed_mentions=AllowedMentions.none(),
-        )
-
-        await prediction_message.reply("Prediction closed!")
-        await interaction.response.send_message("Prediction closed!", ephemeral=True)
-
-    @app_commands.command(name="payout_prediction")
-    @app_commands.checks.has_role(MOD_ROLE)
-    @app_commands.describe(option="Option to payout")
-    async def payout_prediction(
-        self, interaction: Interaction, option: PredictionChoice
-    ):
-        """Payout predicton to option left or right"""
-        await PayoutPredictionController.payout_prediction(
-            option, interaction, self.client
-        )
-
-    @app_commands.command(name="redo_payout")
-    @app_commands.checks.has_role(MOD_ROLE)
-    @app_commands.describe(option="Option to payout")
-    async def redo_payout(self, interaction: Interaction, option: PredictionOutcome):
-        """Redo the last prediction's payout"""
-        await PayoutPredictionController.redo_payout(option, interaction, self.client)
 
     @app_commands.command(name="set_chat_mode")
     @app_commands.checks.has_role(MOD_ROLE)
